+++ app/src/main/java/eu/kanade/tachiyomi/data/backup/models/Backup.kt
@@ -18,4 +18,8 @@ data class Backup(
     // Global Popular/Latest feeds
     @ProtoNumber(610) var backupFeeds: List<BackupFeed> = emptyList(),
     // KMK <--
+    
+    // Chimahon -->
+    @ProtoNumber(700) var backupNovels: List<BackupNovel> = emptyList(),
+    // Chimahon <--
 )